tests/wpt/tests/encoding-detection/pl-ISO-8859-2-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>pl ISO-8859-2 late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "ISO-8859-2", 'Expected ISO-8859-2');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/pl-ISO-8859-2-late.sub.html");
</script>
